requirements.txt
Django==4.2.7
Pillow==10.1.0
reportlab==4.0.7
razorpay==1.3.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
django-extensions==3.2.3
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0